Introduce generic fillers specified by a function tag
The new approach stores which function is used to calculate the variable
to fill the histogram in the filler as a function pointer. The various
functions are defined as members of the root filler class. Which
particular function is used is selected based on a user supplied const
char* tag which is used to locate the function in a static std::map of
function tag to function pointer. This will help with the proliferation
of filler subclasses which needed to be defined for each variable that
needs to be filled. Now one needs to only define a function in the root
filler class that calculates a variable to be filled based on a set of
TLorentzVectors which are passed to it as const references. Any
individual filler will only need a pointer to one such function which is
fixed at construction.