Too many good changes (last commit before getting rid of old cruft)
The ghist is now all fillers of 1d and 2d are unified under a single
class, which uses a char tag to figure out what function in the base
class to use when filling the histogram. By supplying multiple
constructors for these unified classes, one can handle all cases (pair
variables, invariants etc...). This lowers the maitenance burden by a
lot since a modification doesn't have to be propagated to all the
independent variable fillers.

This is last commit with all the old stuff. Next commit they will be gone