Unify particle list name and titles into one array of arrays
This allows to pass the information to the constructors using one
variable. In addition to being easier on the fingers, it insures that
the names and the titles correspond. (as demonstrated by a bug in the
previous version where pi+ and pi- were switched between the two arrays
and the name and title did not match)